+++ compiler/builtins.rs
@@ -1,57 +1,13 @@
-use crate::{Object, Objects};
+mod console;
+pub use console::*;
+
+pub mod numbers;
+
+use crate::{Environment, Object};
 use sflynlang_parser::{
     ast::{DataType, DataTypes},
     Error, Position,
 };
-
-pub fn print(
-    objects: Vec<Object>,
-    position: Position,
-) -> Result<Object, Error> {
-    if objects.len() != 1 {
-        return Err(Error::new_expect_arguments(position, 1, objects.len()));
-    }
-
-    let object = objects[0].clone();
-
-    if object.get_node().get_string().is_none() {
-        return Err(Error::new_expect_type(
-            object.get_position(),
-            "String",
-            &object.to_data_type().to_string(),
-        ));
-    }
-
-    println!("{}", object.to_string());
-
-    Ok(Object::new(object.get_position(), Objects::Void))
-}
-
-pub fn debug(
-    objects: Vec<Object>,
-    position: Position,
-    debug_mode: bool,
-) -> Result<Object, Error> {
-    if objects.len() != 1 {
-        return Err(Error::new_expect_arguments(position, 1, objects.len()));
-    }
-
-    let object = objects[0].clone();
-
-    if object.get_node().get_string().is_none() {
-        return Err(Error::new_expect_type(
-            object.get_position(),
-            "String",
-            &object.to_data_type().to_string(),
-        ));
-    }
-
-    if debug_mode {
-        println!("{}", object.to_string());
-    }
-
-    Ok(Object::new(object.get_position(), Objects::Void))
-}
 
 pub fn check_builtin(
     key: String,
@@ -80,5 +36,20 @@ pub fn check_builtin(
         }
 
         _ => Err(Error::new_unknown_identifier(position, key)),
+    }
+}
+
+pub fn eval_builtin(
+    key: String,
+    position: Position,
+    arguments: Vec<Object>,
+    environment: &mut Environment,
+) -> Result<Object, Error> {
+    match key.as_str() {
+        "print" => console::print(arguments, position),
+        "debug" => {
+            console::debug(arguments, position, environment.is_debug_mode())
+        }
+        _ => Err(Error::new_unknown_token(position)),
     }
 }