+++ compiler/environment/store.rs
@@ -2,7 +2,7 @@ use crate::Object;
 use sflynlang_parser::ast::DataType;
 use std::collections::HashMap;
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub struct Store {
     data_types: HashMap<String, DataType>,
     objects: HashMap<String, Object>,
@@ -50,6 +50,24 @@ impl Store {
                 && self.get_outer().unwrap().has_data_type_with_outer(key))
     }
 
+    pub fn get_data_type(&self, key: &String) -> Option<DataType> {
+        if self.has_data_type(key) {
+            Some(self.get_data_types().get(key).unwrap().clone())
+        } else {
+            None
+        }
+    }
+
+    pub fn get_data_type_with_outer(&self, key: &String) -> Option<DataType> {
+        match self.get_data_type(key) {
+            Some(data_type) => Some(data_type),
+            None => match self.get_outer() {
+                Some(outer) => outer.get_data_type_with_outer(key),
+                None => None,
+            },
+        }
+    }
+
     pub fn has_object(&self, key: &String) -> bool {
         self.objects.contains_key(key)
     }
@@ -60,21 +78,41 @@ impl Store {
                 && self.get_outer().unwrap().has_object_with_outer(key))
     }
 
+    pub fn get_object(&self, key: &String) -> Option<Object> {
+        if self.has_object(key) {
+            Some(self.get_objects().get(key).unwrap().clone())
+        } else {
+            None
+        }
+    }
+
+    pub fn get_object_with_outer(&self, key: &String) -> Option<Object> {
+        match self.get_object(key) {
+            Some(object) => Some(object),
+            None => match self.get_outer() {
+                Some(outer) => outer.get_object_with_outer(key),
+                None => None,
+            },
+        }
+    }
+
     pub fn is_builtin(&self, key: &String) -> bool {
         key == "print" || key == "debug"
     }
 
-    pub fn has_key(&self, key: &String) -> bool {
-        self.is_builtin(key)
-            || self.has_data_type_with_outer(key)
-            || self.has_object_with_outer(key)
+    pub fn has_key_type(&self, key: &String) -> bool {
+        self.is_builtin(key) || self.has_data_type_with_outer(key)
     }
 
-    pub fn add_data_type(&mut self, key: String, value: DataType) {
-        self.data_types.insert(key, value);
+    pub fn has_key_object(&self, key: &String) -> bool {
+        self.is_builtin(key) || self.has_object_with_outer(key)
     }
 
-    pub fn add_object(&mut self, key: String, value: Object) {
-        self.objects.insert(key, value);
+    pub fn add_data_type(&mut self, key: &String, value: &DataType) {
+        self.data_types.insert(key.clone(), value.clone());
+    }
+
+    pub fn add_object(&mut self, key: &String, value: &Object) {
+        self.objects.insert(key.clone(), value.clone());
     }
 }